requirements.txt
Django==2.2.5
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
django-extensions==2.2.1
coreapi==2.3.3
PyYAML==5.1.2
uritemplate==3.0.0